requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
boto3==1.26.144
botocore==1.29.144
build==1.1.1
cachetools==5.3.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
defusedxml==0.7.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.42
gTTS==2.5.0
idna==2.10
importlib-resources==6.1.1
importlib_metadata==7.0.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
labelImg==1.8.6
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==23.2
pandas==2.0.3
parsimonious==0.10.0
pillow==10.2.0
pkgutil_resolve_name==1.3.10
playsound==1.3.0
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
pytorch-nlp==0.5.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.19
rpds-py==0.18.0
s3transfer==0.6.1
scipy==1.10.0
seaborn==0.13.1
simple-image-download==0.4
singlestoredb==1.0.2
six==1.16.0
smmap==5.0.1
sqlparams==6.0.1
streamlit==1.32.2
supervision==0.18.0
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.4
urllib3==2.1.0
watchdog==4.0.0
zipp==3.17.0
